refactor(survey): extract structure-sync to keep survey-db under the size ratchet
The survey question-edit fix grew survey-db.server.ts to 1085 lines, past its
928-line allowlist baseline (tools:check-file-size, CI quality job). The ratchet
is meant to move down, not grow allowlisted offenders, so extract the added
concern rather than bump the baseline.

updateSurveyMetadata + syncSurveyStructure moved verbatim to a focused
survey-structure.server.ts; updateSurveyMetadata is re-exported from
survey-db.server.ts so every existing importer (the surveys action + both tests)
is unchanged. survey-db.server.ts is back to 927 lines. typecheck, lint,
check-file-size, and the survey suites all pass.
